Day-08-AI-Todo-CLI/requirements.txt
click
google-generativeai
python-dotenv
rich
